+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: `ContributionCard`: Update styles.
+- Fix: `VerticalProgress`: Update styles.
 
 ## [3.23.0] - 2021-07-06